implemented auth bloc in ui
